[#207] fix(api): lint import
