+++ kafka-ui-api/src/main/java/com/provectus/kafka/ui/strategy/ksql/statement/ExplainStrategy.java
@@ -1,17 +1,17 @@
-package com.provectus.kafka.ui.strategy.ksqlStatement;
+package com.provectus.kafka.ui.strategy.ksql.statement;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.provectus.kafka.ui.model.KsqlCommandResponse;
 import org.springframework.stereotype.Component;
 
 @Component
-public class TerminateStrategy extends KsqlStatementStrategy {
+public class ExplainStrategy extends KsqlStatementStrategy {
   private final String requestPath = "/ksql";
-  private final String responseValueKey = "commandStatus";
+  private final String responseValueKey = "queryDescription";
 
   @Override
   public KsqlCommandResponse serializeResponse(JsonNode response) {
-    return serializeMessageResponse(response, responseValueKey);
+    return serializeTableResponse(response, responseValueKey);
   }
 
   @Override
@@ -21,6 +21,6 @@ public class TerminateStrategy extends KsqlStatementStrategy {
 
   @Override
   protected String getTestRegExp() {
-    return "terminate (.*);";
+    return "explain (.*);";
   }
 }
