+++ kafka-ui-api/src/main/java/com/provectus/kafka/ui/strategy/ksql/statement/ListStrategy.java
@@ -1,18 +1,16 @@
-package com.provectus.kafka.ui.strategy.ksqlStatement;
+package com.provectus.kafka.ui.strategy.ksql.statement;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.provectus.kafka.ui.model.KsqlCommandResponse;
-import org.springframework.stereotype.Component;
-
 import java.util.List;
 import java.util.Optional;
+import org.springframework.stereotype.Component;
 
 @Component
-public class ShowStrategy extends KsqlStatementStrategy {
+public class ListStrategy extends KsqlStatementStrategy {
   private final String requestPath = "/ksql";
+  private final List<String> statements = List.of("functions", "topics", "streams", "tables");
   private String responseValueKey = "";
-  private final List<String> statements =
-      List.of("functions", "topics", "streams", "tables", "queries", "properties");
 
   @Override
   public KsqlCommandResponse serializeResponse(JsonNode response) {
@@ -42,11 +40,10 @@ public class ShowStrategy extends KsqlStatementStrategy {
   }
 
   private String getTestRegExp(String key) {
-    return "show " + key + ";";
+    return "list " + key + ";";
   }
 
   private void setResponseValueKey(String path) {
     responseValueKey = path;
   }
-
 }
