+++ kafka-ui-api/src/test/java/com/provectus/kafka/ui/strategy/ksql/statement/DropStrategyTest.java
@@ -1,4 +1,9 @@
-package com.provectus.kafka.ui.strategy.ksqlStatement;
+package com.provectus.kafka.ui.strategy.ksql.statement;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -10,17 +15,14 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
-
 @ExtendWith(MockitoExtension.class)
-class CreateStrategyTest {
+class DropStrategyTest {
+  private final ObjectMapper mapper = new ObjectMapper();
   private KsqlStatementStrategy ksqlStatementStrategy;
-  private ObjectMapper mapper = new ObjectMapper();
 
   @BeforeEach
   public void setUp() {
-    ksqlStatementStrategy = new CreateStrategy();
+    ksqlStatementStrategy = new DropStrategy();
   }
 
   @Test
@@ -31,26 +33,16 @@ class CreateStrategyTest {
 
   @Test
   public void shouldReturnTrueInTest() {
-    assertTrue(ksqlStatementStrategy.test("CREATE STREAM stream WITH (KAFKA_TOPIC='topic');"));
-    assertTrue(ksqlStatementStrategy.test("CREATE STREAM stream" +
-        " AS SELECT users.id AS userid FROM users EMIT CHANGES;"
-    ));
-    assertTrue(ksqlStatementStrategy.test(
-        "CREATE TABLE table (id VARCHAR) WITH (KAFKA_TOPIC='table');"
-    ));
-    assertTrue(ksqlStatementStrategy.test(
-        "CREATE TABLE pageviews_regions WITH (KEY_FORMAT='JSON')" +
-            "  AS SELECT gender, COUNT(*) AS numbers" +
-            "  FROM pageviews EMIT CHANGES;"
-    ));
+    assertTrue(ksqlStatementStrategy.test("drop table table1;"));
+    assertTrue(ksqlStatementStrategy.test("drop stream stream2;"));
   }
 
   @Test
   public void shouldReturnFalseInTest() {
     assertFalse(ksqlStatementStrategy.test("show streams;"));
     assertFalse(ksqlStatementStrategy.test("show tables;"));
-    assertFalse(ksqlStatementStrategy.test("CREATE TABLE test;"));
-    assertFalse(ksqlStatementStrategy.test("CREATE STREAM test;"));
+    assertFalse(ksqlStatementStrategy.test("create table test;"));
+    assertFalse(ksqlStatementStrategy.test("create stream test;"));
   }
 
   @Test
