+++ kafka-ui-api/src/test/java/com/provectus/kafka/ui/strategy/ksql/statement/ListStrategyTest.java
@@ -1,31 +1,30 @@
-package com.provectus.kafka.ui.strategy.ksqlStatement;
+package com.provectus.kafka.ui.strategy.ksql.statement;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.provectus.kafka.ui.exception.UnprocessableEntityException;
-import com.provectus.kafka.ui.model.KsqlCommand;
 import com.provectus.kafka.ui.model.KsqlCommandResponse;
-import com.provectus.kafka.ui.model.KsqlResponseTable;
 import com.provectus.kafka.ui.model.Table;
+import java.util.List;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
-
 @ExtendWith(MockitoExtension.class)
-class ShowStrategyTest {
+class ListStrategyTest {
+  private final ObjectMapper mapper = new ObjectMapper();
   private KsqlStatementStrategy ksqlStatementStrategy;
-  private ObjectMapper mapper = new ObjectMapper();
 
   @BeforeEach
   public void setUp() {
-    ksqlStatementStrategy = new ShowStrategy();
+    ksqlStatementStrategy = new ListStrategy();
   }
 
   @Test
@@ -36,28 +35,24 @@ class ShowStrategyTest {
 
   @Test
   public void shouldReturnTrueInTest() {
-    assertTrue(ksqlStatementStrategy.test("SHOW STREAMS;"));
-    assertTrue(ksqlStatementStrategy.test("SHOW TABLES;"));
-    assertTrue(ksqlStatementStrategy.test("SHOW TOPICS;"));
-    assertTrue(ksqlStatementStrategy.test("SHOW QUERIES;"));
-    assertTrue(ksqlStatementStrategy.test("SHOW PROPERTIES;"));
-    assertTrue(ksqlStatementStrategy.test("SHOW FUNCTIONS;"));
+    assertTrue(ksqlStatementStrategy.test("LIST STREAMS;"));
+    assertTrue(ksqlStatementStrategy.test("LIST TABLES;"));
+    assertTrue(ksqlStatementStrategy.test("LIST TOPICS;"));
+    assertTrue(ksqlStatementStrategy.test("LIST FUNCTIONS;"));
   }
 
   @Test
   public void shouldReturnFalseInTest() {
-    assertFalse(ksqlStatementStrategy.test("LIST STREAMS;"));
-    assertFalse(ksqlStatementStrategy.test("LIST TABLES;"));
-    assertFalse(ksqlStatementStrategy.test("LIST TOPICS;"));
-    assertFalse(ksqlStatementStrategy.test("LIST QUERIES;"));
-    assertFalse(ksqlStatementStrategy.test("LIST PROPERTIES;"));
-    assertFalse(ksqlStatementStrategy.test("LIST FUNCTIONS;"));
+    assertFalse(ksqlStatementStrategy.test("SHOW STREAMS;"));
+    assertFalse(ksqlStatementStrategy.test("SHOW TABLES;"));
+    assertFalse(ksqlStatementStrategy.test("SHOW TOPICS;"));
+    assertFalse(ksqlStatementStrategy.test("SHOW FUNCTIONS;"));
   }
 
   @Test
   public void shouldSerializeStreamsResponse() {
     JsonNode node = getResponseWithData("streams");
-    ksqlStatementStrategy.test("show streams;");
+    ksqlStatementStrategy.test("list streams;");
     KsqlCommandResponse serializedResponse = ksqlStatementStrategy.serializeResponse(node);
     Table table = serializedResponse.getData();
     assertThat(table.getHeaders()).isEqualTo(List.of("header"));
@@ -67,7 +62,7 @@ class ShowStrategyTest {
   @Test
   public void shouldSerializeTablesResponse() {
     JsonNode node = getResponseWithData("tables");
-    ksqlStatementStrategy.test("show tables;");
+    ksqlStatementStrategy.test("list tables;");
     KsqlCommandResponse serializedResponse = ksqlStatementStrategy.serializeResponse(node);
     Table table = serializedResponse.getData();
     assertThat(table.getHeaders()).isEqualTo(List.of("header"));
@@ -77,17 +72,7 @@ class ShowStrategyTest {
   @Test
   public void shouldSerializeTopicsResponse() {
     JsonNode node = getResponseWithData("topics");
-    ksqlStatementStrategy.test("show topics;");
-    KsqlCommandResponse serializedResponse = ksqlStatementStrategy.serializeResponse(node);
-    Table table = serializedResponse.getData();
-    assertThat(table.getHeaders()).isEqualTo(List.of("header"));
-    assertThat(table.getRows()).isEqualTo(List.of(List.of("value")));
-  }
-
-  @Test
-  public void shouldSerializePropertiesResponse() {
-    JsonNode node = getResponseWithData("properties");
-    ksqlStatementStrategy.test("show properties;");
+    ksqlStatementStrategy.test("list topics;");
     KsqlCommandResponse serializedResponse = ksqlStatementStrategy.serializeResponse(node);
     Table table = serializedResponse.getData();
     assertThat(table.getHeaders()).isEqualTo(List.of("header"));
@@ -97,17 +82,7 @@ class ShowStrategyTest {
   @Test
   public void shouldSerializeFunctionsResponse() {
     JsonNode node = getResponseWithData("functions");
-    ksqlStatementStrategy.test("show functions;");
-    KsqlCommandResponse serializedResponse = ksqlStatementStrategy.serializeResponse(node);
-    Table table = serializedResponse.getData();
-    assertThat(table.getHeaders()).isEqualTo(List.of("header"));
-    assertThat(table.getRows()).isEqualTo(List.of(List.of("value")));
-  }
-
-  @Test
-  public void shouldSerializeQueriesResponse() {
-    JsonNode node = getResponseWithData("queries");
-    ksqlStatementStrategy.test("show queries;");
+    ksqlStatementStrategy.test("list functions;");
     KsqlCommandResponse serializedResponse = ksqlStatementStrategy.serializeResponse(node);
     Table table = serializedResponse.getData();
     assertThat(table.getHeaders()).isEqualTo(List.of("header"));
@@ -117,7 +92,7 @@ class ShowStrategyTest {
   @Test
   public void shouldSerializeWithException() {
     JsonNode node = getResponseWithData("streams");
-    ksqlStatementStrategy.test("show tables;");
+    ksqlStatementStrategy.test("list tables;");
     Exception exception = assertThrows(
         UnprocessableEntityException.class,
         () -> ksqlStatementStrategy.serializeResponse(node)
